requirements.txt
# Snowflake Demo Requirements for Pharmacy2U
# Python 3.9+ required

# Snowflake connectivity
snowflake-connector-python==3.12.0
snowflake-snowpark-python==1.20.0

# Data generation and processing
pandas==2.2.0
numpy==1.26.0
faker==22.0.0
Faker-Healthcare==1.1.3

# Data visualization
plotly==5.18.0
altair==5.2.0
streamlit==1.39.0

# Machine learning
scikit-learn==1.4.0
xgboost==2.0.3

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0

# Development tools
pytest==8.0.0
black==24.1.0
ruff==0.2.0

# Snowflake CLI
snowflake-cli-labs==2.0.0
